requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
discord.py>=2.3.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
httpx>=0.25.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pillow>=10.1.0
prometheus-client>=0.19.0
google-genai>=1.0.0
